Large update: P200 to OSC path, multi thread status, batch script
The python server now has a status thread which protects the
communication to the camera with a mutex (uneeded now...). The server
has code to check the camera responses for errors ect, and retry if a
command didnt get through. But its buggy and so it is currently
disabled.
The status thread asks the camera for its focus mode (auto or manual)
and then updates an "led" in the open stage control panel. This can
later be expanded to show a panel of where the camera is currently
pointing and ts current zoom, but this seems low priority as there are
previow monitors to actually see what the camera is looking at.

Open stage control now looks for the python server on local host:
127.0.0.1 instead of fixed IP, this makes it much m ore portable.

Lastly, I added a "run" script which starts the python server and runs
the open stage control will some useful settings.

TODO: Have this run script use a config file to determine which midi
device to use. I may never get to this TODO item...